move mischooks code to separate file
